more router fixing
